Added start menu
Added a basic start menu that now shows up when the game is booted up. Two buttons exist so far: Start, which takes you to Level 1, and Quit, which quits the game. Also input mappings have been tweaked to allow WASD and all gamepads can be used to control the player.