Added level timer
Added a timer that records how quickly the player completes the level.  Plus, the Level Completed screen now includes a button to retry and a button to go to the next level instead of automatically going to the next level. Also removed an unnecessary reference in the start menu and made sure all classes specify the type of variables they are using.